chainparams: move CCheckpointData into chainparams.h
This unties CChainParams from its dependency on checkpoints. Instead, now it
only depends on the raw checkpoint data.

Conflicts:
	src/test/Checkpoints_tests.cpp
